Orders

+++ internal/model/order.go
@@ -4,7 +4,7 @@ type (
 	Order struct {
 		Number     int     `json:"number,string"`
 		Status     string  `json:"status"`
-		Accrual    float32 `json:"accrual,omitempty"`
+		Accrual    float64 `json:"accrual,omitempty"`
 		UploadedAt string  `json:"uploaded_at"`
 	}
 	OrderInPoll struct {
@@ -15,6 +15,6 @@ type (
 	OrderInAccrual struct {
 		Number  int     `json:"order,string"`
 		Status  string  `json:"status"`
-		Accrual float32 `json:"accrual,omitempty"`
+		Accrual float64 `json:"accrual,omitempty"`
 	}
 )
